requirements.txt
pandas
faker
python-dotenv
requests
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary
pydantic-sqlalchemy==0.0.9